+++ tests/conftest.py
@@ -3,7 +3,7 @@ import pytest
 
 from ingenialink.canopen.network import CanopenNetwork, CAN_DEVICE, CAN_BAUDRATE
 from ingenialink.ethernet.network import EthernetNetwork
-from tests.virtual_drive import VirtualDrive
+from ingenialink.virtual.virtual_drive import VirtualDrive
 from ingenialink.eoe.network import EoENetwork
 
 ALLOW_PROTOCOLS = ["no_connection", "ethernet", "ethercat", "canopen", "eoe"]
@@ -55,7 +55,6 @@ def connect_canopen(protocol_contents):
     servo = net.connect_to_slave(
         target=protocol_contents["node_id"],
         dictionary=protocol_contents["dictionary"],
-        eds=protocol_contents["eds"],
     )
     return servo, net
 
@@ -101,7 +100,7 @@ def connect_to_slave(pytestconfig, read_config):
 def virtual_drive():
     test_ip = "127.0.0.1"
     test_port = 81
-    server = VirtualDrive(test_ip, test_port)
+    server = VirtualDrive(test_ip, test_port, dictionary_path="./tests/resources/virtual_drive.xdf")
     server.start()
     yield server
     server.stop()